GitHub issue#1462, Code refactor to load cleaner JSON files instead of repeated JS functions + Moved common function to index.js + Imported in all the json configs
