+++ src/config/projectQuestions/generic_chatbot.v1.0.js
@@ -1,17 +1,6 @@
 import _ from 'lodash'
 // import { Icons } from 'appirio-tech-react-components'
-import { findProduct} from '../projectWizard'
-
-const isFileRequired = (project, subSections) => {
-  const subSection = _.find(subSections, (s) => s.type === 'questions')
-  const fields = _.filter(subSection.questions, q => q.type.indexOf('see-attached') > -1)
-  // iterate over all seeAttached type fields to check
-  //  if any see attached is checked.
-  return _.some(_.map(
-    _.map(fields, 'fieldName'),
-    fn => _.get(project, `${fn}.seeAttached`)
-  ))
-}
+import { findProduct, isFileRequired } from '../projectWizard'
 
 const sections = [
   {
