+++ src/config/projectQuestions/topcoder.v1.js
@@ -1,16 +1,6 @@
 import _ from 'lodash'
 import { Icons } from 'appirio-tech-react-components'
-
-const isFileRequired = (project, subSections) => {
-  const subSection = _.find(subSections, (s) => s.type === 'questions')
-  const fields = _.filter(subSection.questions, q => q.type.indexOf('see-attached') > -1)
-  // iterate over all seeAttached type fields to check
-  //  if any see attached is checked.
-  return _.some(_.map(
-    _.map(fields, 'fieldName'),
-    fn => _.get(project, `${fn}.seeAttached`)
-  ))
-}
+import { isFileRequired } from '../projectWizard'
 
 const sections = [
   {
