+++ src/config/projectWizard/index.js
@@ -441,3 +441,23 @@ export function getProjectCreationTemplateField(product, sectionId, subSectionId
   }
   return null
 }
+
+/**
+ * Finds if files are required for project
+ *
+ * @param {object} project       project object
+ * @param {object} subSections   subSections object
+ *
+ * @return {boolean} true if files required, else false
+ */
+
+export function isFileRequired(project, subSections) {
+  const subSection = _.find(subSections, (s) => s.type === 'questions')
+  const fields = _.filter(subSection.questions, q => q.type.indexOf('see-attached') > -1)
+  // iterate over all seeAttached type fields to check
+  //  if any see attached is checked.
+  return _.some(_.map(
+    _.map(fields, 'fieldName'),
+    fn => _.get(project, `${fn}.seeAttached`)
+  ))
+}
